[css-anchor-position-1] Fix anchor-center alignment corner case https://bugs.webkit.org/show_bug.cgi?id=284619
Reviewed by NOBODY (OOPS!).

This change fixes the anchor-center behavior when anchor target has width: auto.

As anchor positioning (in case of anchor-center) forces inset properties to
resolve to 0, in case of width: auto, the wrong code path was triggered and
effectively width was calculated to take maximum possible size.

* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-center-003-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-center-003.html: Added.
* Source/WebCore/rendering/RenderBox.cpp:
(WebCore::RenderBox::computePositionedLogicalWidth const):