[GTK][WPE] Integrate MALLOC_HEAP_BREAKDOWN with sysprof https://bugs.webkit.org/show_bug.cgi?id=300430
Reviewed by NOBODY (OOPS!).

This change integrates MALLOC_HEAP_BREAKDOWN feature with sysprof on
GTK and WPE ports.

With this change, when the sysprof is actively used to profile WebKit,
the heap usage breakdown is reported to sysprof using sysprof counters
instead of printing to stdout.

Other than providing the integration itself, this change also
introduces a couple workarounds for the libsysprof-capture library
bug that results in generating colliding counter IDs across threads.